style: moved all modules to be base level
