+++ src/inventory-item/inventory-item.controller.spec.ts
@@ -1,11 +1,11 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { InventoryItemController } from './inventory-item.controller';
 import { InventoryItem } from './inventory-item.entity';
-import { Inventory } from '../inventory.entity';
+import { Inventory } from '../inventory/inventory.entity';
 import { getRepositoryToken } from '@nestjs/typeorm';
 import { InventoryItemService } from './inventory-item.service';
-import { Model } from './model/model.entity';
-import { Part } from './part/part.entity';
+import { Model } from '../model/model.entity';
+import { Part } from '../part/part.entity';
 
 export const mockRepository = jest.fn(() => ({
   metadata: {
