+++ src/inventory-item/inventory-item.module.ts
@@ -3,9 +3,9 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { InventoryItemController } from './inventory-item.controller';
 import { InventoryItemService } from './inventory-item.service';
 import { InventoryItem } from './inventory-item.entity';
-import { Inventory } from '../inventory.entity';
-import { Model } from './model/model.entity';
-import { Part } from './part/part.entity';
+import { Inventory } from '../inventory/inventory.entity';
+import { Model } from '../model/model.entity';
+import { Part } from '../part/part.entity';
 
 @Module({
   imports: [TypeOrmModule.forFeature([InventoryItem, Inventory, Model, Part])],
