--- src/inventory/dto/create-inventory.dto.ts
@@ -1,4 +1,0 @@
-export class CreateInventoryDto {
-  // This should define the properties of the dto that is passed with the request when creating an inventory.
-  // Fill out later.
-}
