+++ src/listing/listing.module.ts
@@ -4,7 +4,7 @@ import { ListingController } from './listing.controller';
 import { ListingService } from './listing.service';
 import { Listing } from './listing.entity';
 import { User } from '../user/user.entity';
-import { InventoryItem } from '../inventory/inventory-item/inventory-item.entity';
+import { InventoryItem } from '../inventory-item/inventory-item.entity';
 
 @Module({
   imports: [TypeOrmModule.forFeature([Listing, User, InventoryItem])],
