+++ src/listing/listing.service.spec.ts
@@ -3,7 +3,7 @@ import { ListingService } from './listing.service';
 import { getRepositoryToken } from '@nestjs/typeorm';
 import { Listing } from './listing.entity';
 import { User } from '../user/user.entity';
-import { InventoryItem } from '../inventory/inventory-item/inventory-item.entity';
+import { InventoryItem } from '../inventory-item/inventory-item.entity';
 
 export const mockRepository = jest.fn(() => ({
   metadata: {
