+++ src/message/message.entity.ts
@@ -1,5 +1,5 @@
-import { Conversation } from '../../conversation/conversation.entity';
-import { User } from '../../user/user.entity';
+import { Conversation } from '../conversation/conversation.entity';
+import { User } from '../user/user.entity';
 import {
   Entity,
   Column,
